feat: :sparkles: Add product selection, dynamic options, and summary features with invoice generation.
The product component was updated to add the option to select products based on the saved items in the 'product' file in data. Additionally, dynamic options were generated according to the number of products saved in the product file. An event listener was added to the product name select field so that, when a user selects a product, the unit price and product code are auto-filled (these values cannot be changed by the user). The "Add" button functionality was enhanced to trigger a custom event containing an object with key/value pairs for the code, name, unit price, quantity, and subtotal. If a valid product or quantity is not selected, an alert message prompts the user to enter the required fields.  A summary component was created, which includes a table where products entered by the user are added. This table updates every time a product is added, using the 'details' data from the custom event triggered in the products component.  Furthermore, a detail invoice component was created, which displays a card showing the subtotal of the products, VAT (tax amount), and total after tax. These values update dynamically whenever a new product is added to the table. The "Pay" button generates an invoice by collecting the header data, products from the summary table, and the subtotal, VAT, and total values, then saves them in local storage with a defined data structure, cleans the fields, and generates a new invoice number for future use.  Required imports were added in app.js, and necessary custom elements were included in index.html.